requirements.txt
discord.py
requests
openai
web3
openai-agents
pinecone
aiohttp 
python-dotenv
mcp
uvicorn